requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-aiplatform>=1.38.0
gspread
python-docx
tenacity
python-dotenv
fpdf
requests